+++ src/modes/survey/IndicatorRevisions.svelte
@@ -3,7 +3,7 @@
   import IndicatorRevision from './IndicatorRevision.svelte';
 
   /**
-   * @type {import('./questions').Question}
+   * @type {import('../../stores/questions').Question}
    */
   export let question;
   /**
